Update Offer And Stripe Controller

+++ app/Http/Controllers/OfferController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\OfferRequest;
+use App\Models\Flight;
 use App\Models\Log;
 use App\Models\Offer;
 use Illuminate\Http\Request;
@@ -19,6 +20,17 @@ class OfferController extends Controller
             $path = $offerRequest->file('image')->storePublicly('OfferImage', 'public');
         }
 
+        $flight = Flight::find($offerRequest->flight_id);
+
+        if ($flight->days == '[]') {
+            return error('some thing went wrong', 'this flight have not any schedule day', 422);
+        }
+        foreach ($flight->offers as $offer) {
+            if (($offer->start_date <= $offerRequest->start_date && $offer->end_date >= $offerRequest->start_date) || ($offer->end_date >= $offerRequest->end_date && $offer->end_date <= $offerRequest->start_date)) {
+                return error('some thing went wrong', 'there is an offer in this date', 422);
+            }
+        }
+
         $offer = Offer::create([
             'employee_id' => Auth::guard('user')->user()->employee->employee_id,
             'flight_id' => $offerRequest->flight_id,
@@ -27,6 +39,7 @@ class OfferController extends Controller
             'end_date' => $offerRequest->end_date,
             'image' => 'storage/' . $path,
             'title' => $offerRequest->title,
+            'discount' => $offerRequest->discount,
         ]);
 
         Log::create([
@@ -41,6 +54,17 @@ class OfferController extends Controller
     public function updateOffer(Offer $offer, OfferRequest $offerRequest)
     {
         $user = Auth::guard('user')->user();
+
+        $flight = Flight::find($offerRequest->flight_id);
+
+        if ($flight->days == '[]') {
+            return error('some thing went wrong', 'this flight have not any schedule day', 422);
+        }
+        foreach ($flight->offers as $off) {
+            if (($off->start_date <= $offerRequest->start_date && $off->end_date >= $offerRequest->start_date) || ($off->end_date >= $offerRequest->end_date && $off->end_date <= $offerRequest->start_date)) {
+                return error('some thing went wrong', 'there is an offer in this date', 422);
+            }
+        }
         if ($offerRequest->file('image')) {
             if (File::exists($offer->image)) {
                 File::delete($offer->image);
@@ -58,6 +82,7 @@ class OfferController extends Controller
             'start_date' => $offerRequest->start_date,
             'end_date' => $offerRequest->end_date,
             'title' => $offerRequest->title,
+            'discount' => $offerRequest->discount,
         ]);
 
         Log::create([
